requirements.txt
streamlit
groq
sentence-transformers
faiss-cpu
numpy
python-dotenv
pandas
pypdf2
jira
pytest
fastapi
uvicorn
python-multipart
requests
